Revert "jump_label: Fix sparc64 warning"
This reverts commit eb7fbe28f0bb70cd15e78b9b72a70e61a56ae6ad.

+++ kernel/jump_label.c
@@ -374,8 +374,7 @@ static void __jump_label_update(struct static_key *key,
 			if (kernel_text_address(entry->code))
 				arch_jump_label_transform(entry, jump_label_type(entry));
 			else
-				WARN_ONCE(1, "can't patch jump_label at %pS",
-					  (void *)(unsigned long)entry->code);
+				WARN_ONCE(1, "can't patch jump_label at %pS", (void *)entry->code);
 		}
 	}
 }
